Use variant-title to compile ToC: https://github.com/metanorma/metanorma-bipm/issues/148
